refactor(tauri): extract the sidecar lifecycle into sidecar.rs
Unit 11e of the lib.rs split. 1075 lines move verbatim -- spawn with
health-poll and exponential backoff, the graceful /api/shutdown stop and
hard-kill fallback, `restart_sidecar`, the port and exe-lock waits the
updater consumes, the two spawn-env path resolvers, the Windows
port-holder diagnostic, and `port_holder_tests`. 21 lines differ from
their originals and every one is the `pub(crate)` the move requires.
`lib.rs` 5101 -> 4036.

The boundary came from the source, not from the epic's line-range guess.
`SIDECAR_HEALTHY` and `SIDECAR_GAVE_UP` read like sidecar state and
deliberately stayed: both are documented as read and written ONLY under
the `PendingOpens` mutex, their writers are the queue helpers, and their
test module is `pending_opens_tests`. Splitting a flag from the mutex
that supplies its ordering proof is the hazard `SIDECAR_GAVE_UP`'s own
doc comment warns about.

Privacy running downward is not name resolution, and conflating them is
how a move like this fails to compile. A private crate-root item IS
visible to a descendant module, but a bare path written in `lib.rs` does
not resolve there, so eleven crate-root names needed importing -- plus
`Emitter` and `Manager`, which appear nowhere in the moved text because
they are trait imports that `.emit(` and the ten `.state()` calls need
for method resolution, and whose absence is an E0599 rather than an
unresolved name.

`sidecar_job` is Windows-gated as a MODULE, so an ungated
`use crate::sidecar_job;` is an unconditional E0432 on macOS and Linux
while compiling clean here; the import mirrors the cfg, which keeps the
one call site byte-identical. `system_paths` is already fully qualified
at its call site and needed nothing. Neither is checkable locally --
`cargo check --target x86_64-unknown-linux-gnu` dies in `glib-sys` for
want of Linux pkg-config -- so CI is the only oracle for that class.

The verbatim check runs in BOTH directions, and the reverse half is not
ceremony: a duplicate left behind in `lib.rs` compiles, because the two
copies sit in different modules and nothing in the crate or `ci.yml`
denies warnings. Mutant N1 confirmed it -- survived every gate with a
single `constant SHUTDOWN_URL is never used`.

`url_constants_tests` stays in `lib.rs` and imports the four constants
that left. It pins the whole loopback-URL family, including `OPEN_URL`
and both `LAUNCHER_*_URL`s, so moving it would have widened three
constants outside this cluster for a test's sake -- the same call 11a
made for `check_for_update_now`.
