+++ docs/architecture.md
@@ -371,7 +371,7 @@ The bootstrap turn is likewise written **on spawn**, not on the CLI's `init` lin
 
 ### Activation
 
-The channel shim is registered **only on explicit request** — `tandem setup --apply --with-channel-shim`. There is deliberately **no wizard checkbox** — the CLI flag is the only opt-in, and three places claimed otherwise until 2026-08-09; the wizard's apply route calls `shouldRegisterChannelShim` with no override and so can only ever write the HTTP entry. It does not *remove* an existing shim either, because its `remove` list comes from the user's confirmed diff. It was default-on for the Claude Code target from #985 until Track E (see [Why `tandem-channel` Is Opt-In](#why-tandem-channel-is-opt-in) for why that default was doing harm). On the desktop bundle the wizard's apply path uses the channel-shim path injected into the sidecar via `TANDEM_CHANNEL_DIST` on spawn (resolved from the resource dir by `src-tauri/src/lib.rs`), since the server's own package-root derivation resolves outside the bundle.
+The channel shim is registered **only on explicit request** — `tandem setup --apply --with-channel-shim`. There is deliberately **no wizard checkbox** — the CLI flag is the only opt-in, and three places claimed otherwise until 2026-08-09; the wizard's apply route calls `shouldRegisterChannelShim` with no override and so can only ever write the HTTP entry. It does not *remove* an existing shim either, because its `remove` list comes from the user's confirmed diff. It was default-on for the Claude Code target from #985 until Track E (see [Why `tandem-channel` Is Opt-In](#why-tandem-channel-is-opt-in) for why that default was doing harm). On the desktop bundle the wizard's apply path uses the channel-shim path injected into the sidecar via `TANDEM_CHANNEL_DIST` on spawn (resolved from the resource dir by `src-tauri/src/sidecar.rs`), since the server's own package-root derivation resolves outside the bundle.
 
 **Registration is not activation.** Registering the shim gets the subprocess spawned; whether Claude Code then *honors* the channel is decided separately, inside Claude Code, and is never reported back. Activation requires a **hand-started interactive session** launched with the dev-channels flag:
 
@@ -1002,7 +1002,7 @@ The flagless alternative to the channel shim, run as `tandem monitor` by the plu
 - `tauri.conf.json` -- App config: identifier (`com.tandem.editor`), window dimensions (1200×800, min 800×600), `bundle.externalBin` (**two**: `binaries/node-sidecar` and `binaries/tandem-reaper` — both must exist or `cargo tauri dev/build` fails its existence check), `bundle.resources` (`dist/server/`, `dist/channel/`, `dist/stdio-bridge/`, `dist/client/`, `sample/`, `skills/`, `CHANGELOG.md`, `docs/workflows.md` — `tauri_build::build()` checks this list at build time, so keep it in sync here), CSP, updater endpoint (GitHub Releases `latest.json`), `bundle.createUpdaterArtifacts: true`
 - `capabilities/default.json` -- Core permissions: window events, shell sidecar, fs read/write, dialog
 - `capabilities/desktop.json` -- Desktop-only permissions: single-instance, window-state save/restore, updater
-- `src/lib.rs` -- The Tauri entry logic (not all of it — the sibling modules below carry a growing share, and Unit 11 of the maintainability programme is still moving clusters out, so this list is authoritative and any count you add to it will not be): plugin registration (single-instance **first**), sidecar lifecycle (spawn, health-poll, exponential backoff, kill on exit), `resolve_channel_dist()` (injects `TANDEM_CHANNEL_DIST` into the sidecar so the channel shim registers from the resource dir; replaced `run_setup()`/`/api/setup` in #477 PR 3c-ii-c), system tray build + event handlers (tray "Setup AI Assistant" emits `open-integration-wizard`), window hide-on-close, auto-updater (launch + periodic 8h), `strip_win_prefix()` for Windows `\\?\` paths, `copy_sample_files()` (first-run copy to app-data dir)
+- `src/lib.rs` -- The Tauri entry logic (not all of it — the sibling modules below carry a growing share, and Unit 11 of the maintainability programme is still moving clusters out, so this list is authoritative and any count you add to it will not be): plugin registration (single-instance **first**), system tray build + event handlers (tray "Setup AI Assistant" emits `open-integration-wizard`), window hide-on-close, auto-updater (launch + periodic 8h), `strip_win_prefix()` for Windows `\\?\` paths, `copy_sample_files()` (first-run copy to app-data dir)
 - `src/main.rs` -- Entry point, delegates to `lib::run()`
 - `src/autostart.rs` -- Start-at-login registration (ADR-046, #1236)
 - `src/bounded_command.rs`, `src/single_flight.rs` -- Deadline-bounded external process spawns, and the in-flight guard that keeps a slow probe from piling up once it is off the main thread (#1371)
@@ -1017,6 +1017,7 @@ The flagless alternative to the channel shim, run as `tandem monitor` by the plu
 - `src/integrations_probe.rs` -- Detects installed AI clients for the wizard
 - `src/keychain.rs`, `src/token_store.rs` -- OS keychain access and auth-token storage
 - `src/sentry_reporting.rs` -- Opt-in crash reporting (`TANDEM_SENTRY_DSN`)
+- `src/sidecar.rs` -- The Node sidecar's process lifecycle (Unit 11e): spawn with health-poll and exponential backoff, graceful `/api/shutdown` stop then hard kill, `restart_sidecar`, the port/exe-lock waits the updater uses, `resolve_channel_dist()` / `resolve_stdio_bridge_dist()` (which inject `TANDEM_CHANNEL_DIST` / `TANDEM_STDIO_BRIDGE_DIST` so the shim and the Claude Desktop stdio entry resolve from the resource dir; replaced `run_setup()`/`/api/setup` in #477 PR 3c-ii-c), and the Windows port-holder diagnostic that names what is squatting :3478/:3479. `SIDECAR_HEALTHY` and the pending-opens queue stay in `lib.rs` -- they are read and written under the `PendingOpens` mutex, not by this module
 - `src/sidecar_job.rs` -- Windows job-object containment for the sidecar
 - `src/uninstall_scrub.rs` -- The scrub the NSIS uninstaller hook invokes (Cowork entries, firewall rules, start-at-login). Distinct from `src/cli/uninstall-scrub.ts`, which is the npm CLI's scrub (MCP entries + skill) and is never invoked by NSIS.
 - `src/win_app_mode.rs` -- Windows app-mode detection
