+++ docs/decisions.md
@@ -1139,7 +1139,7 @@ Everything above renders identically. Recovering any of it needs a per-node sour
 
 **Status:** Accepted (2026-06-07)
 
-**Context:** The Tauri auto-updater path (`perform_install` in `src-tauri/src/lib.rs`) runs `kill_sidecar` → `wait_for_port_release` (+ Windows `wait_for_sidecar_unlock`) → `update.download_and_install` (minisign-verified) → `app.restart()`. Issue #925 flagged two gaps: (1) no rollback if `app.restart()` fails to relaunch a broken/corrupt new binary, and (2) no health-poll *inside* the updater path after restart.
+**Context:** The Tauri auto-updater path (`perform_install` in `src-tauri/src/lib.rs`) runs `kill_sidecar` → `wait_for_port_release` (+ Windows `wait_for_sidecar_unlock`; all three moved to `src-tauri/src/sidecar.rs` in Unit 11e) → `update.download_and_install` (minisign-verified) → `app.restart()`. Issue #925 flagged two gaps: (1) no rollback if `app.restart()` fails to relaunch a broken/corrupt new binary, and (2) no health-poll *inside* the updater path after restart.
 
 **Decision:** Ship v1 with current behavior unchanged (option **(c)** of #925). The decisive constraint is that **Tauri v2 `AppHandle::restart()` is divergent** — it exits and relaunches the process without returning (tauri-apps/tauri #12310/#13923/#11392). Therefore: (2) an in-`perform_install` post-restart probe is unreachable dead code by construction; and (1) rollback cannot be driven from the old process, which is gone the instant `restart()` runs — true rollback would require a standalone watchdog/bootstrapper process plus a `.previous` binary copy and platform-specific swap-back logic. Neither the [Tauri v2 updater docs](https://v2.tauri.app/plugin/updater/) nor [CrabNebula's auto-updates guide](https://docs.crabnebula.dev/guides/auto-updates-tauri) recommend either; both stop at `download_and_install()` + relaunch with no rollback or post-restart verification.
